moved mastr db classes into separate file
